extract code to searchrequestlogger and trim down waitOnSearchComplete calls
