+++ tests/acceptance/acceptancesuites/facetsonload.js
@@ -4,7 +4,6 @@ import {
   FACETS_ON_LOAD_PAGE
 } from '../server';
 import FacetsPage from '../pageobjects/facetspage';
-import VerticalPage from '../pageobjects/verticalpage';
 import { RequestLogger } from 'testcafe';
 import {
   browserBackButton,
@@ -13,32 +12,27 @@ import {
   registerIE11NoCacheHook
 } from '../utils';
 import { getMostRecentQueryParamsFromLogger } from '../requestUtils';
+import { VERTICAL_SEARCH_URL_REGEX } from '../constants';
 
 fixture`Facets page`
-  .beforeEach(async t => {
-    await VerticalPage.getVerticalResultsComponent().registerLogger(t);
-  })
   .before(setupServer)
   .after(shutdownServer)
   .page`${FACETS_ON_LOAD_PAGE}`;
 
 test('Facets work with back/forward navigation and page refresh', async t => {
-  const searchQueryUrl = VerticalPage.getVerticalResultsComponent().getSearchQueryUrl();
   const logger = RequestLogger({
-    url: searchQueryUrl
+    url: VERTICAL_SEARCH_URL_REGEX
   });
   await t.addRequestHooks(logger);
-  await registerIE11NoCacheHook(t, searchQueryUrl);
+  await registerIE11NoCacheHook(t, VERTICAL_SEARCH_URL_REGEX);
 
   async function getFacetsFromRequest () {
     const urlParams = await getMostRecentQueryParamsFromLogger(logger);
     return JSON.parse(urlParams.get('facetFilters'));
   }
   const searchComponent = FacetsPage.getSearchComponent();
-  const resultComponent = VerticalPage.getVerticalResultsComponent();
   await searchComponent.enterQuery('all');
   await searchComponent.submitQuery();
-  await resultComponent.waitOnSearchComplete();
 
   let currentFacets = await getFacetsFromRequest();
   await t.expect(currentFacets).eql({});
