+++ tests/acceptance/acceptancesuites/filterboxsuite.js
@@ -4,7 +4,6 @@ import {
   FILTERBOX_PAGE
 } from '../server';
 import FacetsPage from '../pageobjects/facetspage';
-import VerticalPage from '../pageobjects/verticalpage';
 import { Selector, RequestLogger } from 'testcafe';
 import {
   browserBackButton,
@@ -17,27 +16,27 @@ import {
   expectRequestLocationRadiusToEql,
   expectRequestDoesNotContainParam
 } from '../requestUtils';
+import { VERTICAL_SEARCH_URL_REGEX } from '../constants';
+import SearchRequestLogger from '../searchrequestlogger';
 
 fixture`FilterBox page`
   .beforeEach(async t => {
-    await VerticalPage.getVerticalResultsComponent().registerLogger(t);
+    await SearchRequestLogger.registerVerticalSearchLogger(t);
   })
   .before(setupServer)
   .after(shutdownServer)
   .page`${FILTERBOX_PAGE}`;
 
 test('single option filterbox works with back/forward navigation and page refresh', async t => {
-  const searchQueryUrl = VerticalPage.getVerticalResultsComponent().getSearchQueryUrl();
   const radiusFilterLogger = RequestLogger({
-    url: searchQueryUrl
+    url: VERTICAL_SEARCH_URL_REGEX
   });
   await t.addRequestHooks(radiusFilterLogger);
-  await registerIE11NoCacheHook(t, searchQueryUrl);
+  await registerIE11NoCacheHook(t, VERTICAL_SEARCH_URL_REGEX);
   const searchComponent = FacetsPage.getSearchComponent();
-  const resultComponent = VerticalPage.getVerticalResultsComponent();
   await searchComponent.enterQuery('all');
   await searchComponent.submitQuery();
-  await resultComponent.waitOnSearchComplete();
+  await SearchRequestLogger.waitOnSearchComplete();
 
   const filterBox = FacetsPage.getStaticFilterBox();
   const radiusFilter = await filterBox.getFilterOptions('DISTANCE');
@@ -45,39 +44,38 @@ test('single option filterbox works with back/forward navigation and page refres
   // Choose the 25 miles radius filter
   await radiusFilter.toggleOption('25 miles');
   const filterTags = Selector('.yxt-ResultsHeader-removableFilterTag');
-  await resultComponent.waitOnSearchComplete();
+  await SearchRequestLogger.waitOnSearchComplete();
   await t.expect(filterTags.count).eql(1);
   await expectRequestLocationRadiusToEql(radiusFilterLogger, 40233.6);
   radiusFilterLogger.clear();
 
   // Hit the back button
   await browserBackButton();
-  await resultComponent.waitOnSearchComplete();
+  await SearchRequestLogger.waitOnSearchComplete();
   await t.expect(filterTags.count).eql(0);
   await expectRequestDoesNotContainParam(radiusFilterLogger, 'locationRadius');
   radiusFilterLogger.clear();
 
   // Hit the forward button
   await browserForwardButton();
-  await resultComponent.waitOnSearchComplete();
+  await SearchRequestLogger.waitOnSearchComplete();
   await t.expect(filterTags.count).eql(1);
   await expectRequestLocationRadiusToEql(radiusFilterLogger, 40233.6);
   radiusFilterLogger.clear();
 
   // Refresh the page
   await browserRefreshPage();
-  await resultComponent.waitOnSearchComplete();
+  await SearchRequestLogger.waitOnSearchComplete();
   await t.expect(filterTags.count).eql(1);
   await expectRequestLocationRadiusToEql(radiusFilterLogger, 40233.6);
 });
 
 test('multioption filterbox works with back/forward navigation and page refresh', async t => {
-  const searchQueryUrl = VerticalPage.getVerticalResultsComponent().getSearchQueryUrl();
   const filterBoxLogger = RequestLogger({
-    url: searchQueryUrl
+    url: VERTICAL_SEARCH_URL_REGEX
   });
   await t.addRequestHooks(filterBoxLogger);
-  await registerIE11NoCacheHook(t, searchQueryUrl);
+  await registerIE11NoCacheHook(t, VERTICAL_SEARCH_URL_REGEX);
   const martyFilter = {
     c_puppyPreference: {
       $eq: 'Marty'
@@ -92,21 +90,19 @@ test('multioption filterbox works with back/forward navigation and page refresh'
   const staticFilter = await filterBox.getFilterOptions('STATIC FILTERS');
   const filterTags = Selector('.yxt-ResultsHeader-removableFilterTag');
   const searchComponent = FacetsPage.getSearchComponent();
-  const resultComponent = VerticalPage.getVerticalResultsComponent();
   await searchComponent.enterQuery('all');
   await searchComponent.submitQuery();
-  await resultComponent.waitOnSearchComplete();
 
   // Click the 'Marty' checkbox
   await staticFilter.toggleOption('Marty');
-  await resultComponent.waitOnSearchComplete();
+  await SearchRequestLogger.waitOnSearchComplete();
   await t.expect(await getRequestFilters(filterBoxLogger)).eql(martyFilter);
   await t.expect(filterTags.count).eql(1);
   filterBoxLogger.clear();
 
   // Click the 'Frodo' checkbox
   await staticFilter.toggleOption('Frodo');
-  await resultComponent.waitOnSearchComplete();
+  await SearchRequestLogger.waitOnSearchComplete();
   await t.expect(filterTags.count).eql(2);
   await t.expect(await getRequestFilters(filterBoxLogger)).eql({
     $or: [martyFilter, frodoFilter]
@@ -115,27 +111,27 @@ test('multioption filterbox works with back/forward navigation and page refresh'
 
   // Hit the back button, see the 'Frodo' filter disappear
   await browserBackButton();
-  await resultComponent.waitOnSearchComplete();
+  await SearchRequestLogger.waitOnSearchComplete();
   await t.expect(await getRequestFilters(filterBoxLogger)).eql(martyFilter);
   await t.expect(filterTags.count).eql(1);
   filterBoxLogger.clear();
 
   // Hit the back button, see the 'Marty' filter disappear
   await browserBackButton();
-  await resultComponent.waitOnSearchComplete();
+  await SearchRequestLogger.waitOnSearchComplete();
   await expectRequestDoesNotContainParam(filterBoxLogger, 'filters');
   await t.expect(filterTags.count).eql(0);
 
   // Hit the forward button, see the 'Marty' filter applied again
   await browserForwardButton();
-  await resultComponent.waitOnSearchComplete();
+  await SearchRequestLogger.waitOnSearchComplete();
   await t.expect(await getRequestFilters(filterBoxLogger)).eql(martyFilter);
   await t.expect(filterTags.count).eql(1);
   filterBoxLogger.clear();
 
   // Hit the forward button, see the 'Frodo' filter applied again
   await browserForwardButton();
-  await resultComponent.waitOnSearchComplete();
+  await SearchRequestLogger.waitOnSearchComplete();
   await t.expect(filterTags.count).eql(2);
   await t.expect(await getRequestFilters(filterBoxLogger)).eql({
     $or: [martyFilter, frodoFilter]
@@ -144,7 +140,7 @@ test('multioption filterbox works with back/forward navigation and page refresh'
 
   // Refresh the page
   await browserRefreshPage();
-  await resultComponent.waitOnSearchComplete();
+  await SearchRequestLogger.waitOnSearchComplete();
   await t.expect(filterTags.count).eql(2);
   await t.expect(await getRequestFilters(filterBoxLogger)).eql({
     $or: [martyFilter, frodoFilter]
@@ -152,16 +148,15 @@ test('multioption filterbox works with back/forward navigation and page refresh'
 });
 
 test('locationRadius of 0 is persisted', async t => {
-  const resultComponent = VerticalPage.getVerticalResultsComponent();
   const filterBox = FacetsPage.getStaticFilterBox();
   const radiusFilter = await filterBox.getFilterOptions('DISTANCE');
   await radiusFilter.toggleOption('0 miles');
   const filterTags = Selector('.yxt-ResultsHeader-removableFilterTag');
-  await resultComponent.waitOnSearchComplete();
+  await SearchRequestLogger.waitOnSearchComplete();
   await t.expect(filterTags.count).eql(1);
 
   // Refresh the page
   await browserRefreshPage();
-  await resultComponent.waitOnSearchComplete();
+  await SearchRequestLogger.waitOnSearchComplete();
   await t.expect(filterTags.count).eql(1);
 });
