+++ tests/acceptance/constants.js
@@ -1,0 +1,2 @@
+module.exports.VERTICAL_SEARCH_URL_REGEX = /v2\/accounts\/me\/answers\/vertical\/query/;
+module.exports.UNIVERSAL_SEARCH_URL_REGEX = /v2\/accounts\/me\/answers\/query/;
